Update strongbox-authorization-yaml.md
Same issue with this page as well. None of the links are working. After checking multiple other pages, none of the external links were working for me. I am unsure if it is my computer and connection or the wiki. All of the links that correspond to other pages on the wiki are working fine.