Update getting-started-with-persistence.md
The http://127.0.0.1:2480/studio/index.html link does not work. I am assuming because the link would become active once the OrientDB has been installed and configured. Should we maybe state (this link will not work until OrientDB has been installed and configured). I am not familiar with Orient and do not know if this location might vary depending on the develoeprs system.

In the Creating a DAO layer, I changed the semicolons for periods as each line was a complete sentence.