Update logging.md
The logging.md file that I have in my system is the one that I corrected. I noticed just before sending this that the file name that is in the repo was add-logging.md. Please let me know if the changes to this file look good. If they do, I can update the add-logging.md file seperately in a different commit.