Update maven-indexer.md
lines 86-89 are not formated properly. In the code, it is formated to have a new line for each of the points but in the wiki it does not express that new line and instead just runs them all together. I do not know enough about markup to adjust this. If anyone knows how, please let me know and I will correct it in a new commit. Thank you.